Decoder: Use serde_bencode to decode bencoded data
Remove my implementation of a bencode decoder and use serde_bencode
instead. My implementation relied on the data being a valid utf-8
String which was not the case, also relying on the serde_json to
represent data was a mistake since serde_json does not have a way to
represent raw bytes (as far as I can tell).